+++ lib/urldata.h
@@ -20,7 +20,7 @@
  * This software is distributed on an "AS IS" basis, WITHOUT WARRANTY OF ANY
  * KIND, either express or implied.
  *
- * $Id: urldata.h,v 1.195 2004-02-22 22:31:24 bagder Exp $
+ * $Id: urldata.h,v 1.196 2004-02-26 13:40:43 bagder Exp $
  ***************************************************************************/
 
 /* This file is for lib internal stuff */
@@ -205,7 +205,7 @@ struct negotiatedata {
  ***************************************************************************/
 struct HTTP {
   struct FormData *sendit;
-  int postsize;
+  size_t postsize;
   char *postdata;
 
   const char *p_pragma;      /* Pragma: string */
